// WvDatasourceManage/src/models/WvDsManageStore.ts
export default class WvDsManageStore {
    library:  Array<Object>= new Array<Object>();
    libraryVersion: number = 1;
    datasourceId: string = null;
    apiRootUrl:string = "";
    pageDatasourceId: string = null;
    pageDatasourceName: string = null;
    pageDatasourceParams: Array<Object> = new Array<Object>();
 }